requirements.txt
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
gtts==2.3.2
playsound==1.3.0
Pillow==10.1.0
numpy==1.24.3
accelerate==0.24.1